用户与提交信息

+++ zscacm/src/main/java/com/example/zscacm/entity/CfContests.java
@@ -23,7 +23,7 @@ public class CfContests {
 
     private Date beginTime;
 
-    private int relativeTime;
+    private long relativeTime;
 
 
 }
